+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Editing:
     * New quantization algorithm available (Edge Exclusion Modes)
+* UI:
+    * Simplified version of the trace dialog
 
 ## 0.5.0